Update to Minecraft 1.8

+++ src/main/java/net/blay09/mods/eiramoticons/render/EmoticonBuffer.java
@@ -6,8 +6,8 @@ package net.blay09.mods.eiramoticons.render;
 import net.blay09.mods.eiramoticons.emoticon.Emoticon;
 
 public class EmoticonBuffer {
-	public int[] positionX = new int[0];
-	public int[] positionY = new int[0];
+	public float[] positionX = new float[0];
+	public float[] positionY = new float[0];
 	public float[] alpha = new float[0];
 	public Emoticon[] emoticons = new Emoticon[0];
 }
